requirements.txt
#tensorflow-gpu==1.8.0
bleach==1.5.0
cycler==0.10.0
decorator==4.1.2
html5lib==0.9999999
Markdown==2.6.9
matplotlib==2.1.0
networkx==2.0
nltk==3.2.5
numpy==1.13.3
olefile==0.44
pandas==0.21.0
Pillow==4.3.0
protobuf==3.4.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyWavelets==0.5.2
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
tensorflow-tensorboard==0.1.8
test-tube==0.502
Werkzeug==0.12.2